Move imported into separate file
This makes it easier to refactor it.

Also moving requires into a single place - the main cloudfront-rails file.
